fix(timestamp): stamp records with true int64 epoch nanoseconds
Root cause of every Parquet event_date collapsing to 1970-01-01: the daemon
stamped float64(startPollTime.UnixNano())/1e9 — fractional epoch SECONDS
(~1.75e9) — onto a field named/typed as nanoseconds. The derived event_date
(time.Unix(0, ns)) then read ~1.75e9 as ns → ~1.7s past epoch → 1970-01-01
for every row. The /1e9 was load-bearing only for ClickHouse, which reads a
numeric DateTime64 input as seconds; it broke Parquet/Snowflake and the
humanized CSV/JSON timestamps.

Standardize on true nanoseconds as a single int64, faithful to the source
(the poll-time time.Now()):

- proto: XtcpFlatRecord.timestamp_ns double -> int64 (regenerated Go/vtproto/
  cpp/dart/python + ClickHouse format schema via `nix run .#regen-protos`).
  Breaking wire change (fixed64 -> varint); safe because daemon, Kafka schema,
  and consumers roll out together in the coordinated release.
- deserialize.go: stamp startPollTime.UnixNano() verbatim (drop /1e9).
- s3parquet: ParquetRow.TimestampNs int64; utcDateFromNs(int64); drop the
  now-redundant int64() cast in the event_date memoization.
- recordfmt.TimestampRFC3339(int64): true-ns arithmetic.
- ClickHouse DDL: kafka-engine table timestamp_ns DateTime64(9) -> Int64 (raw
  ns passthrough); MV converts with fromUnixTimestamp64Nano before landing in
  the MergeTree table (which stays DateTime64(9)); same for the _testing table.
  The converted column is emitted first with `* EXCEPT (timestamp_ns)` so the
  positional SELECT->target mapping is preserved.
- docs/parquet-format.md: timestamp_ns column type double -> int64.

All unit tests (both dest_s3parquet and default tags) pass; the Commit-1
failing tests now go green.

+++ pkg/xtcp/destinations_s3parquet.go
@@ -411,7 +411,7 @@ func (d *s3ParquetDest) worker(ctx context.Context) {
 			parquetRow := rowFromProto(row)
 			// Stamp the derived event_date, re-formatting only when the UTC day
 			// changes (≈once per file).
-			if idx := int64(row.TimestampNs) / nsPerDayCst; idx != lastDayIdx {
+			if idx := row.TimestampNs / nsPerDayCst; idx != lastDayIdx {
 				lastDayIdx = idx
 				lastDate = utcDateFromNs(row.TimestampNs)
 			}
@@ -735,8 +735,8 @@ func approxRowBytes(r *xtcp_flat_record.XtcpFlatRecord) int {
 }
 
 // utcDateFromNs formats an epoch-ns timestamp as its UTC date (YYYY-MM-DD).
-func utcDateFromNs(ns float64) string {
-	return time.Unix(0, int64(ns)).UTC().Format("2006-01-02")
+func utcDateFromNs(ns int64) string {
+	return time.Unix(0, ns).UTC().Format("2006-01-02")
 }
 
 // rowFromProto translates one *xtcp_flat_record.XtcpFlatRecord into a
